Extract countdown to its own component
