+++ packages/lib-auth/src/views/oauth/components/TimeoutWarningPopup/locales/en.json
@@ -1,5 +1,6 @@
 {
   "TimeoutWarningPopup": {
-    "message": "Your login session will expire soon. To extend your session, click the button."
+    "message": "Your login session will expire soon. To extend your session, click the button.",
+    "button": "Stay logged in"
   }
 }
