+++ packages/lib-auth/src/views/oauth/components/shared/Countdown/index.js
@@ -1,0 +1,1 @@
+export { default } from './Countdown'
